refactor: move Input element to a separate component
